Issue 49618 - Increase cachememsize and dncachememsize variable sizes
Description: nssapld-cachememsize is reverted to a smaller value
if the cap is reached. Increase it to UINT64.

https://pagure.io/389-ds-base/issue/49618

Reviewed by: tbordaz, mreynolds (Thanks!)

+++ ldap/servers/slapd/back-ldbm/ldbm_config.c
@@ -2118,7 +2118,7 @@ ldbm_config_get(void *arg, config_info *config, char *buf)
         break;
     case CONFIG_TYPE_SIZE_T:
         val = (size_t)config->config_get_fn(arg);
-        sprintf(buf, "%lu", (long unsigned int)val);
+        sprintf(buf, "%" PRIu32, (uint32_t)val);
         break;
     case CONFIG_TYPE_UINT64:
         val = (uint64_t)((uintptr_t)config->config_get_fn(arg));
